pyproject.toml
[project]
name = "finance-agent-from-scratch"
version = "0.1.0"
description = "A financial AI agent built from scratch with pure Python, no LangChain or LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "edgartools>=5.0.0",
    "fastapi>=0.119.1",
    "groq>=0.33.0",
    "hdbscan>=0.8.40",
    "instructor>=1.13.0",
    "langfuse>=3.11.2",
    "psycopg[binary]>=3.2",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.1.1",
    "qdrant-client[fastembed]>=1.15.1",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0",
    "torch>=2.11.0",
    "trafilatura>=2.0.0",
    "transformers>=4.57.1",
    "uvicorn>=0.38.0",
    "yfinance>=0.2.66",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
]
